Moved the data collection logic to the Plutograph component. Fixed and optimised it
